Eject phased paulis, split two qubit out
