+++ cirq/optimizers/merge_interactions_test.py
@@ -169,8 +169,8 @@ def test_decompose_partial_czs(circuit):
     cz_gates = [op.gate for op in circuit.all_operations()
                 if isinstance(op, cirq.GateOperation) and
                 isinstance(op.gate, cirq.CZPowGate)]
-    num_full_cz = sum(1 for cz in cz_gates if cz.exponent == 1)
-    num_part_cz = sum(1 for cz in cz_gates if cz.exponent != 1)
+    num_full_cz = sum(1 for cz in cz_gates if cz.exponent % 2 == 1)
+    num_part_cz = sum(1 for cz in cz_gates if cz.exponent % 2 != 1)
     assert num_full_cz == 2
     assert num_part_cz == 0
 
@@ -186,8 +186,8 @@ def test_not_decompose_partial_czs():
     cz_gates = [op.gate for op in circuit.all_operations()
                 if isinstance(op, cirq.GateOperation) and
                 isinstance(op.gate, cirq.CZPowGate)]
-    num_full_cz = sum(1 for cz in cz_gates if cz.exponent == 1)
-    num_part_cz = sum(1 for cz in cz_gates if cz.exponent != 1)
+    num_full_cz = sum(1 for cz in cz_gates if cz.exponent % 2  == 1)
+    num_part_cz = sum(1 for cz in cz_gates if cz.exponent % 2 != 1)
     assert num_full_cz == 0
     assert num_part_cz == 1
 
@@ -213,6 +213,7 @@ def test_post_clean_up():
     optimizer = cirq.MergeInteractions(allow_partial_czs=False,
                                        post_clean_up=clean_up)
     optimizer.optimize_circuit(circuit)
+    cirq.DropEmptyMoments().optimize_circuit(circuit)
 
     assert isinstance(circuit[0].operations[0].gate, Marker)
     assert isinstance(circuit[-1].operations[0].gate, Marker)
